Remove pointless cast.
git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@34517 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/demux_ogg.c
@@ -1431,7 +1431,7 @@ static void demux_ogg_seek(demuxer_t *demuxer, float rel_seek_secs,
         ds         = demuxer->audio;
         os         = &ogg_d->subs[ds->id];
         vi         = &(os->vi);
-        rate       = (float)vi->rate;
+        rate       = vi->rate;
         samplesize = ((sh_audio_t*)ds->sh)->samplesize;
     }
